workerstats moved to worker
