+++ dimos/core/resource_monitor/__init__.py
@@ -3,7 +3,8 @@ from dimos.core.resource_monitor.logger import (
     ResourceLogger,
     StructlogResourceLogger,
 )
-from dimos.core.resource_monitor.stats import ProcessStats, WorkerStats, collect_process_stats
+from dimos.core.resource_monitor.stats import ProcessStats, collect_process_stats
+from dimos.core.worker import WorkerStats
 
 __all__ = [
     "LCMResourceLogger",
